refactorized expedia scraper apis
